requirement.txt
appdirs==1.4.4
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
Django==4.2
djangorestframework==3.14.0
fake-useragent==1.1.3
feedparser==6.0.10
fonttools==4.39.3
frozendict==2.3.7
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multitasking==0.0.11
mysqlclient==2.1.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parse==1.19.0
Pillow==9.5.0
pycparser==2.21
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-html==0.10.0
scikit-learn==1.2.2
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.18
zipp==3.15.0
